Move L2 contract to contracts package
