+++ safe_cli/contracts/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa F401
+from .safe_to_l2_migration import safe_to_l2_migration
